+++ src/main/java/org/neo4j/spark/Neo4JavaSparkContext.java
@@ -6,6 +6,7 @@ import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
 import org.apache.spark.sql.SQLContext;
+import org.neo4j.spark.dataframe.Neo4jDataFrame;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
