+++ update_notes.md
@@ -1,0 +1,6 @@
+List of updates:
+
+* Moved the classes `Neo4jPartition`, `Neo4jRowRDD` and `Neo4jTupleRDD` under `org.neo4j.spark.rdd`
+* Moved the classes `Neo4jDataFrame` and `Neo4jGraphFrames` under `org.neo4j.spark.dataframe`
+* Extracted `Executor` into a dedicated Object
+* Extracted `Neo4jRDD` into a dedicated Class and moved under `org.neo4j.spark.rdd`
